local/index: copy referenced images to same registry as index
When creating a new local index, if the the referenced images are
not in the same registry as the registry we intend to push the index
we copy the images to the same registry as the index.

NOTE: In future we plan to store the image reference names locally
and only copy the images when push the index to registry.

Signed-off-by: Husni Faiz <ahamedhusni73@gmail.com>